requirements.txt
behave==1.2.6
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.1
cryptography==42.0.2
dbus-python==1.3.2
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.3.0
more-itertools==10.2.0
packaging==23.2
parse==1.20.1
parse-type==0.6.2
pep517==0.13.1
pluggy==1.4.0
pycparser==2.21
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==8.0.0
PyYAML==6.0.1
requests==2.31.0
SecretStorage==3.3.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.0
zipp==3.17.0
